Gen_AI/Task-03/requirements.txt
langchain>=0.2.0
langchain-core>=0.2.0
langchain-community>=0.2.0
langchain-groq>=0.1.4

python-dotenv>=1.0.0

pydantic>=2.0.0
requests>=2.31.0    